Move useCacheQuery to data layer
