+++ assets/src/data/services/apollo/queries/index.ts
@@ -1,0 +1,1 @@
+export { default as useCacheQuery } from './useCacheQuery';
