+++ assets/src/domain/eventEditor/services/apollo/queries/datetimes/useDatetimes.ts
@@ -2,7 +2,7 @@ import { useMemo } from 'react';
 
 import useDatetimeQueryOptions from './useDatetimeQueryOptions';
 import { Datetime, DatetimesList } from '../../types';
-import useCacheQuery from '../useCacheQuery';
+import useCacheQuery from '@dataServices/apollo/queries/useCacheQuery';
 import { getCacheIds } from '@appServices/predicates';
 
 const useDatetimes = (): Array<Datetime> => {
