+++ assets/src/domain/eventEditor/services/apollo/queries/priceTypes/usePriceTypes.ts
@@ -2,7 +2,7 @@ import { useMemo } from 'react';
 
 import usePriceTypeQueryOptions from './usePriceTypeQueryOptions';
 import { PriceType, PriceTypesList } from '../../types';
-import useCacheQuery from '../useCacheQuery';
+import { useCacheQuery } from '@dataServices/apollo/queries';
 import { getCacheIds } from '@appServices/predicates';
 /**
  * A custom react hook for retrieving all the priceTypes from cache
